+++ src/main/java/com/project/sooktoring/auth/enumerate/AuthProvider.java
@@ -1,0 +1,5 @@
+package com.project.sooktoring.auth.enumerate;
+
+public enum AuthProvider {
+    GOOGLE
+}
