+++ src/main/java/com/project/sooktoring/auth/enumerate/Role.java
@@ -1,0 +1,6 @@
+package com.project.sooktoring.auth.enumerate;
+
+public enum Role {
+
+    ROLE_MENTEE, ROLE_MENTOR
+}
