+++ src/main/java/com/project/sooktoring/auth/filter/JwtAuthenticationEntryPoint.java
@@ -1,0 +1,25 @@
+package com.project.sooktoring.auth.filter;
+
+import com.project.sooktoring.common.exception.CustomException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.stereotype.Component;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+import static com.project.sooktoring.common.exception.ErrorCode.INVALID_ACCESS_TOKEN;
+
+@Slf4j
+@Component
+public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
+    @Override
+    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
+        log.error("Responding with unauthorized error. Message = {}", authException.getMessage());
+        throw new CustomException(INVALID_ACCESS_TOKEN);
+    }
+}
