+++ src/main/java/com/project/sooktoring/auth/filter/JwtAuthenticationFilter.java
@@ -1,0 +1,39 @@
+package com.project.sooktoring.auth.filter;
+
+import com.project.sooktoring.auth.jwt.AuthToken;
+import com.project.sooktoring.auth.jwt.AuthTokenProvider;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@RequiredArgsConstructor
+@Component
+public class JwtAuthenticationFilter extends OncePerRequestFilter {
+
+    private final AuthTokenProvider tokenProvider;
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        String authorizationHeader = request.getHeader("Authorization");
+        //프론트에서 보낸 accessToken(백서버에서 발급) 검증
+        if(authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+            String tokenStr = authorizationHeader.substring(7);
+            AuthToken token = tokenProvider.convertAuthToken(tokenStr);
+
+            if(token.validate()) {
+                Authentication authentication = tokenProvider.getAuthentication(token);
+                SecurityContextHolder.getContext().setAuthentication(authentication); //인증된 사용자 정보 Security Context에 저장
+            }
+        }
+
+        filterChain.doFilter(request, response);
+    }
+}
