+++ src/main/java/com/project/sooktoring/auth/info/UserInfo.java
@@ -1,0 +1,8 @@
+package com.project.sooktoring.auth.info;
+
+import com.project.sooktoring.auth.domain.User;
+
+public interface UserInfo {
+
+    User getUser(String idToken);
+}
