+++ src/main/java/com/project/sooktoring/auth/jwt/AuthToken.java
@@ -1,0 +1,85 @@
+package com.project.sooktoring.auth.jwt;
+
+import io.jsonwebtoken.*;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+import java.security.Key;
+import java.util.Date;
+
+@Slf4j
+@Getter
+@RequiredArgsConstructor
+public class AuthToken {
+
+    private final String token;
+    private final Key key; //토큰 생성시 사용할 비밀키
+
+    AuthToken(String providerId, Long userId, Date expiry, Key key) {
+        this.key = key;
+        this.token = createAuthToken(providerId, userId, expiry);
+    }
+
+    AuthToken(Date expiry, Key key) {
+        this.key = key;
+        this.token = createAuthToken(null, null, expiry);
+    }
+
+    private String createAuthToken(String providerId, Long userId, Date expiry) {
+        JwtBuilder jwtBuilder = Jwts.builder()
+                .setSubject(providerId)
+                .signWith(key, SignatureAlgorithm.HS256)
+                .setIssuedAt(new Date())
+                .setExpiration(expiry);
+        return (userId != null ? jwtBuilder.claim("userId", userId) : jwtBuilder).compact();
+    }
+
+    public boolean validate() {
+        return this.getTokenClaims() != null;
+    }
+
+    public Claims getTokenClaims() {
+        try {
+            return Jwts.parserBuilder()
+                    .setSigningKey(key)
+                    .build()
+                    .parseClaimsJws(token)
+                    .getBody();
+        } catch (SecurityException e) {
+            log.error("Invalid JWT signature.");
+        } catch (MalformedJwtException e) {
+            log.error("Invalid JWT token.");
+        } catch (ExpiredJwtException e) {
+            log.error("Expired JWT token.");
+        } catch (UnsupportedJwtException e) {
+            log.error("Unsupported JWT token.");
+        } catch (IllegalArgumentException e) {
+            log.error("JWT token compact of handler are invalid");
+        }
+        return null;
+    }
+
+    //만료된 토큰이어도 Claims 꺼내기 위해
+    public Claims getExpiredTokenClaims() {
+        try {
+            return Jwts.parserBuilder()
+                    .setSigningKey(key)
+                    .build()
+                    .parseClaimsJws(token)
+                    .getBody();
+        } catch (ExpiredJwtException e) {
+            log.error("Expired JWT token.");
+            return e.getClaims();
+        } catch (SecurityException e) {
+            log.error("Invalid JWT signature.");
+        } catch (MalformedJwtException e) {
+            log.error("Invalid JWT token.");
+        } catch (UnsupportedJwtException e) {
+            log.error("Unsupported JWT token.");
+        } catch (IllegalArgumentException e) {
+            log.error("JWT token compact of handler are invalid");
+        }
+        return null;
+    }
+}
