+++ src/main/java/com/project/sooktoring/auth/repository/RefreshTokenRepository.java
@@ -1,0 +1,11 @@
+package com.project.sooktoring.auth.repository;
+
+import com.project.sooktoring.auth.domain.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+
+    Optional<RefreshToken> findByKey(Long key);
+}
