+++ src/main/java/com/project/sooktoring/auth/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.project.sooktoring.auth.repository;
+
+import com.project.sooktoring.auth.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByProviderId(String providerId);
+}
