--- src/main/java/com/project/sooktoring/user/enumerate/AuthProvider.java
@@ -1,5 +1,0 @@
-package com.project.sooktoring.user.enumerate;
-
-public enum AuthProvider {
-    GOOGLE
-}
