--- src/main/java/com/project/sooktoring/user/enumerate/Role.java
@@ -1,6 +1,0 @@
-package com.project.sooktoring.user.enumerate;
-
-public enum Role {
-
-    ROLE_MENTEE, ROLE_MENTOR
-}
