--- src/main/java/com/project/sooktoring/user/info/UserInfo.java
@@ -1,8 +1,0 @@
-package com.project.sooktoring.user.info;
-
-import com.project.sooktoring.user.domain.User;
-
-public interface UserInfo {
-
-    User getUser(String idToken);
-}
