--- src/main/java/com/project/sooktoring/user/repository/RefreshTokenRepository.java
@@ -1,11 +1,0 @@
-package com.project.sooktoring.user.repository;
-
-import com.project.sooktoring.user.domain.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-
-    Optional<RefreshToken> findByKey(Long key);
-}
