--- src/main/java/com/project/sooktoring/user/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.project.sooktoring.user.repository;
-
-import com.project.sooktoring.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByProviderId(String providerId);
-}
